rockchip: resource: Change variable type from 32-bit to 64-bit
(1) For mos, secondary U-Boot use dram addr exceed 4G
    (need 64-bit variable type).
(2) If avb lock, boot.img will be preloaded in dram first to
    get dtb from resource.

+++ arch/arm/include/asm/arch-rockchip/resource_img.h
@@ -15,8 +15,8 @@
 
 struct resource_file {
 	char		name[MAX_FILE_NAME_LEN];
-	uint32_t	blk_start;
-	uint32_t	blk_offset;
+	uint64_t	blk_start;
+	uint64_t	blk_offset;
 	char		hash[MAX_HASH_LEN];
 	uint32_t	hash_size;
 	uint32_t	size;		/* in byte */
